backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
firebase-admin
pydantic
google-genai
langchain
langchain-google-genai
chromadb
fastembed
requests
pillow
qrcode[pil]
